requirements.txt
about-time==4.2.1
alive-progress==3.1.5
grapheme==0.6.0
numpy==1.26.4
pandas==2.2.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
six==1.16.0
tzdata==2024.1
